web-ui/server/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
aiosqlite>=0.19.0

# Configuration
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0

# External APIs
anthropic>=0.18.0
openai>=1.12.0
httpx>=0.26.0

# MCP (same as Python server)
mcp[cli]>=1.4.1
fastmcp>=0.2.0

# Security
cryptography>=42.0.0

# Utilities
python-jose>=3.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
